requirements.txt
aiogram>=3.5,<4
python-dotenv>=1.0,<2
